sched/fair: Fix util_avg of new tasks for asymmetric systems
When a new task wakes-up for the first time, its initial utilization
is set to half of the spare capacity of its CPU. The current
implementation of post_init_entity_util_avg() uses SCHED_CAPACITY_SCALE
directly as a capacity reference. As a result, on a big.LITTLE system, a
new task waking up on an idle little CPU will be given ~512 of util_avg,
even if the CPU's capacity is significantly less than that.

Fix this by computing the spare capacity with arch_scale_cpu_capacity().

Change-Id: I523e6884286a7f34995b5ca4d5913ca0bd856dbe
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>